% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonScrape.R
\name{amazonScrape}
\alias{amazonScrape}
\title{A Function for Scraping Amazon Product Reviews}
\usage{
amazonScrape(product_id = NULL, country = "UK", delay = 5)
}
\arguments{
\item{product_id}{What is the unique, 6 character product ID?}

\item{country}{Should the function run for UK, US or Both? Defaults to UK.}

\item{delay}{How long in seconds should the function pause between pages? Defaults to 5.}
}
\description{
This function scrapes all Amazon product reviews from the UK and/or US sites for a specified product ID.
}
\examples{
amazonScrape("B08SQNY4FS", country = "UK")
}
\keyword{amazon}
