% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazonScrape.R
\name{amazonScrape}
\alias{amazonScrape}
\title{A Function for Amazon Review Scraping}
\usage{
amazonScrape(product_id = NULL, country = "UK", delay = 5)
}
\arguments{
\item{product_id}{What is the unique, 6 character product ID?}

\item{country}{Should the function run for UK, US or Both? Defaults to UK.}

\item{delay}{How long in seconds should the function pause between pages? Defaults to 5.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
amazonScrape()
}
\keyword{amazon}
